refactor(payment): 도메인 서비스 포트 및 인프라 어댑터 금액 타입 Long → BigDecimal 변경
- TossPaymentGateway: confirm/cancel 파라미터 및 내부 record 타입 변경
- TossPaymentGatewayImpl: Toss API 경계에서 longValueExact/BigDecimal.valueOf 변환 유지(toss 요구 타입 Long)
- OrderPaymentValidationGateway/Impl: validate() 파라미터 변경
- 클라이언트 DTO(request/response) 금액 필드 변경
- SellerSettlementPayoutRequestedEventConsumer: 비교 로직 BigDecimal 전환

+++ payment/src/main/java/com/example/payment/domain/service/OrderPaymentValidationItemData.java
@@ -1,10 +1,11 @@
 package com.example.payment.domain.service;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrderPaymentValidationItemData(
         UUID orderItemId,
         UUID sellerId,
-        Long lineAmount
+        BigDecimal lineAmount
 ) {
 }
